TSTF/pip_requirements.txt
absl-py==2.1.0
aiosignal==1.3.1
attrs==23.2.0
backcall==0.2.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
cycler==0.11.0
decorator==5.1.1
distlib==0.3.8
dm-tree==0.1.8
docker-pycreds==0.4.0
exceptiongroup==1.2.2
Farama-Notifications==0.0.4
filelock==3.12.2
fiona==1.9.6
fonttools==4.38.0
frozenlist==1.3.3
geopandas==0.10.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
google-auth-oauthlib==0.4.6
grpcio==1.62.2
gym==0.23.1
gym-notices==0.0.8
gymnasium==0.28.1
idna==3.7
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipython==7.34.0
jax-jumpy==1.0.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.17.3
kiwisolver==1.4.5
lxml==5.2.2
lz4==4.3.2
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
metadrive-simulator==0.4.2.3
msgpack==1.0.5
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.0
Panda3D==1.10.13
panda3d-gltf==0.13
panda3d-simplepbr==0.10
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
pluggy==1.2.0
progressbar==2.5
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pygame==2.6.0
Pygments==2.17.2
pyparsing==3.1.2
pyproj==3.2.1
pyrsistent==0.19.3
pytest==7.4.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.3.0
PyYAML==6.0.1
ray==2.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==2.19.2
setproctitle==1.3.3
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
threadpoolctl==3.1.0
tifffile==2021.11.2
tomli==2.0.1
torch==1.13.1
torch-geometric==2.3.1
tqdm==4.66.4
traitlets==5.9.0
typer==0.12.3
typing_extensions==4.7.1
urllib3==2.0.7
virtualenv==20.26.3
wandb==0.18.7
wcwidth==0.2.13
Werkzeug==2.2.3
yapf==0.40.2
zipp==3.15.0
